chore: refine heuristics enum
