--- src/macaron/malware_analyzer/pypi_heuristics/analysis_result.py
@@ -1,25 +1,0 @@
-# Copyright (c) 2024 - 2024, Oracle and/or its affiliates. All rights reserved.
-# Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl/.
-
-"""This file defines the heuristic-analysis result type enum, and function to modify result."""
-
-from enum import Enum
-
-
-class HeuristicResult(Enum):
-    """Result type indicating the outcome of a heuristic.
-
-    Attributes
-    ----------
-        PASS: Indicates that no suspicious activity was detected.
-        FAIL: Indicates that suspicious activity was detected.
-        SKIP: Indicates that the heuristic check could not be performed due to missing metadata.
-
-    The `SKIP` result occurs when the necessary metadata is not available. This often happens
-    when fetching data through the PyPI API and the relevant data, such as the maintainer's
-    join date or release information, is missing or unavailable.
-    """
-
-    PASS = "PASS"  # nosec
-    FAIL = "FAIL"  # nosec
-    SKIP = "SKIP"  # nosec
