+++ src/macaron/malware_analyzer/pypi_heuristics/base_analyzer.py
@@ -5,8 +5,7 @@
 
 from abc import abstractmethod
 
-from macaron.malware_analyzer.pypi_heuristics.analysis_result import HeuristicResult
-from macaron.malware_analyzer.pypi_heuristics.heuristics import Heuristics
+from macaron.malware_analyzer.pypi_heuristics.heuristics import HeuristicResult, Heuristics
 from macaron.slsa_analyzer.package_registry.pypi_registry import PyPIRegistry
 
 
@@ -21,9 +20,9 @@ class BaseHeuristicAnalyzer:
     ) -> None:
         self.name: str = name
         self.heuristic: Heuristics = heuristic
-        self.depends_on: list[
-            tuple[Heuristics, HeuristicResult]
-        ] | None = depends_on  # Contains the dependent heuristics and the expected result of each heuristic
+        self.depends_on: list[tuple[Heuristics, HeuristicResult]] | None = (
+            depends_on  # Contains the dependent heuristics and the expected result of each heuristic
+        )
 
     @abstractmethod
     def analyze(self, api_client: PyPIRegistry) -> tuple[HeuristicResult, dict]:
